move Source from class to interface in its own module
