+++ packages/components/src/components/Cell.tsx
@@ -1,7 +1,7 @@
 import { asyncRows } from 'hightable'
 import { asyncBufferFromUrl, parquetMetadataAsync } from 'hyparquet'
 import { useEffect, useState } from 'react'
-import { FileSource } from '../lib/filesystem.js'
+import { FileSource } from '../lib/source.js'
 import { parquetDataFrame } from '../lib/tableProvider.js'
 import Breadcrumb, { BreadcrumbConfig } from './Breadcrumb.js'
 import Layout from './Layout.js'
