+++ packages/components/src/components/Folder.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { DirSource, FileMetadata } from '../lib/filesystem.js'
+import { DirSource, FileMetadata } from '../lib/source.js'
 import { cn, formatFileSize, getFileDate, getFileDateShort } from '../lib/utils.js'
 import Breadcrumb, { BreadcrumbConfig } from './Breadcrumb.js'
 import Layout, { Spinner } from './Layout.js'
