+++ packages/components/src/components/viewers/ImageView.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { FileSource } from '../../lib/filesystem.js'
+import { FileSource } from '../../lib/source.js'
 import { contentTypes, parseFileSize } from '../../lib/utils.js'
 import { Spinner } from '../Layout.js'
 import ContentHeader from './ContentHeader.js'
