+++ packages/components/src/components/viewers/MarkdownView.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { FileSource } from '../../lib/filesystem.js'
+import { FileSource } from '../../lib/source.js'
 import { parseFileSize } from '../../lib/utils.js'
 import { Spinner } from '../Layout.js'
 import Markdown from '../Markdown.js'
