+++ packages/components/src/components/viewers/ParquetView.tsx
@@ -1,8 +1,8 @@
 import HighTable, { DataFrame, rowCache } from 'hightable'
 import { asyncBufferFromUrl, parquetMetadataAsync } from 'hyparquet'
 import React, { useCallback, useEffect, useState } from 'react'
-import { FileSource } from '../../lib/filesystem.js'
 import { RoutesConfig, appendSearchParams } from '../../lib/routes.js'
+import { FileSource } from '../../lib/source.js'
 import { parquetDataFrame } from '../../lib/tableProvider.js'
 import { Spinner } from '../Layout.js'
 import CellPanel from './CellPanel.js'
