+++ packages/components/src/components/viewers/Viewer.tsx
@@ -1,4 +1,4 @@
-import { FileSource } from '../../lib/filesystem.js'
+import { FileSource } from '../../lib/source.js'
 import { imageTypes } from '../../lib/utils.js'
 import ImageView from './ImageView.js'
 import MarkdownView from './MarkdownView.js'
