+++ packages/components/src/lib/index.ts
@@ -1,7 +1,7 @@
-export { DirSource, FileSource, FileSystem, HttpFileSystem, HyperparamFileSystem, Source } from './filesystem.js'
-export type { FileKind, FileMetadata, SourcePart } from './filesystem.js'
+export { FileSystem, HttpFileSystem, HyperparamFileSystem } from './filesystem.js'
 export { appendSearchParams, replaceSearchParams } from './routes.js'
 export type { RoutesConfig } from './routes.js'
+export type { DirSource, FileKind, FileMetadata, FileSource, Source, SourcePart } from './source.js'
 export { parquetDataFrame } from './tableProvider.js'
 export { asyncBufferFrom, cn, contentTypes, formatFileSize, getFileDate, getFileDateShort, getFileName, imageTypes, parseFileSize } from './utils.js'
 export { parquetQueryWorker } from './workers/parquetWorkerClient.js'
